main.py
# main.py
from kivy.lang import Builder
from kivy.uix.boxlayout import BoxLayout
from kivymd.app import MDApp
from kivymd.uix.button import MDRaisedButton
from kivymd.uix.label import MDLabel
from kivymd.uix.screen import Screen
from kivymd.uix.screenmanager import ScreenManager
from kivy.uix.image import Image
from kivy.clock import Clock
from pyzbar.pyzbar import decode
from PIL import Image as PILImage
import io

class BarcodeScannerApp(MDApp):
    def build(self):
        self.screen_manager = ScreenManager()
        self.main_screen = Screen(name='main')
        
        layout = BoxLayout(orientation='vertical', padding=10, spacing=10)
        
        self.image = Image()
        layout.add_widget(self.image)
        
        self.result_label = MDLabel(halign='center', theme_text_color='Primary')
        layout.add_widget(self.result_label)
        
        self.scan_button = MDRaisedButton(
            text="Scan Barcode",
            pos_hint={"center_x": 0.5},
            on_release=self.scan_barcode
        )
        layout.add_widget(self.scan_button)
        
        self.main_screen.add_widget(layout)
        self.screen_manager.add_widget(self.main_screen)
        
        return self.screen_manager
    
    def scan_barcode(self, instance):
        # Capture image and decode barcode
        self.capture_image()
    
    def capture_image(self):
        # Simulate image capture (in a real app, integrate with the camera)
        Clock.schedule_once(self.decode_barcode, 1)
    
    def decode_barcode(self, dt):
        # Simulate barcode decoding
        fake_image_data = b''  # Replace with actual image data
        image = PILImage.open(io.BytesIO(fake_image_data))
        decoded_objects = decode(image)
        
        if decoded_objects:
            self.result_label.text = decoded_objects[0].data.decode("utf-8")
        else:
            self.result_label.text = "No barcode detected"

if __name__ == "__main__":
    BarcodeScannerApp().run()

