requirements.txt
appdirs==1.4.4
argcomplete==1.8.1
blinker==1.4
build==1.2.1
buildozer==1.5.0
certifi==2020.6.20
chardet==4.0.0
colorama==0.4.6
command-not-found==0.3
cryptography==3.4.8
Cython==3.0.10
dbus-python==1.2.18
distlib==0.3.8
distro==1.7.0
distro-info==1.1+ubuntu0.2
docutils==0.21.2
filelock==3.14.0
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
jeepney==0.7.1
Jinja2==3.1.4
keyring==23.5.0
Kivy==2.3.0
Kivy-Garden==0.1.5
kivy-garden.xcamera==2020.613
kivy-garden.zbarcam==2020.524
kivy3==0.0.1.dev0
kivymd==1.2.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==2.1.5
more-itertools==8.10.0
netifaces==0.11.0
numpy==1.26.4
oauthlib==3.2.0
opencv-python==4.9.0.80
packaging==24.0
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.2
ptyprocess==0.7.0
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
pyproject_hooks==1.1.0
python-apt==2.4.0+ubuntu3
python-for-android==2024.1.21
PyYAML==5.4.1
pyzbar==0.1.9
requests==2.25.1
sdkmanager==0.5.1
SecretStorage==3.3.1
sh==1.14.3
six==1.16.0
systemd-python==234
toml==0.10.2
tomli==2.0.1
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
virtualenv==20.26.2
wadllib==1.3.6
xcamera==2020.613
zbarcam==2020.524
zipp==1.0.0
